# docs/openapi.json
{
  "components": {
    "responses": {
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Missing or unknown bearer key."
      }
    },
    "schemas": {
...
      "PacksListResponse": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "items": {
              "$ref": "#/components/schemas/AvailablePack"
            },
            "type": "array"
          },
          "installed": {
            "items": {
              "$ref": "#/components/schemas/InstalledPack"
            },
            "type": "array"
          }
        },
        "required": [
          "available",
          "installed"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "API key (`Authorization: Bearer <key>`); its SHA-256 must be registered in BRAIN_API_KEYS. Keys carry scopes (`brain:read`, `brain:write`, `brain:admin`, `brain:read_pii`, `registry:publish`, `registry:curate`, `indexer:write`) — each operation states the scope it requires. Not an OAuth flow.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from the zod wire contracts (`pnpm openapi:build` — scripts/build-openapi.ts); do not edit by hand. Excluded on purpose: the operator/ops surface (jobs, leases, policy, stats, maintenance, GDPR cascades), which is indexed in docs/api.md, and the MCP transport `ALL /mcp/{companyId}`, which is JSON-RPC over Streamable HTTP rather than REST — MCP clients discover it through `tools/list` (docs/mcp.md). Scopes are plain bearer-key grants, not OAuth flows — each operation states its required scope.",
    "license": {
      "identifier": "AGPL-3.0-or-later",
      "name": "AGPL-3.0-or-later"
    },
    "summary": "The community-facing platform surface: the memory core (fact ingest, search, multi-hop, synthesize, entity and fact reads), the Domain Pack registry and tenant pack admin, the document pipeline, the external-indexer work protocol, the raw-substrate driver, the evidence substrate, and source reputation reads.",
    "title": "INITE Brain — Platform API",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/packs": {
      "get": {
        "description": "Builtin packs are globally available and cannot be installed or uninstalled here. Source: src/admin/admin-packs.controller.ts.\n\nRequired scope: `brain:admin`.",
        "operationId": "listDomainPacks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacksListResponse"
                }
              }
            },
            "description": "Available and installed packs."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "List available + installed packs for this tenant",
        "tags": [
          "Domain Packs"
        ]
      },
      "post": {
        "description": "Validates and installs a community / custom manifest, seeding its predicates. `expectedChecksum` pins the exact content.\n\nRequired scope: `brain:admin`.",
        "operationId": "installDomainPack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallPackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallPackResponse"
                }
              }
            },
            "description": "Installed."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "Install a pack manifest into this tenant",
        "tags": [
          "Domain Packs"
        ]
      }
    },
    "/v1/admin/packs/from-registry": {
      "post": {
        "description": "Resolves the manifest (latest non-yanked, or a pinned version) and installs it with the registry checksum pinned — the installed content is exactly what the registry served. A PAID pack without an active `domain_pack:<packId>` entitlement answers 402 with a self-describing hint (the checkout route to call, then retry); while the billing service is unreachable, paid installs fail CLOSED with 503.\n\nRequired scope: `brain:admin`.",
        "operationId": "installDomainPackFromRegistry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallFromRegistryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallPackResponse"
                }
              }
            },
            "description": "Installed."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequiredHint"
                }
              }
            },
            "description": "Paid pack, no entitlement — purchase via the checkout route in the hint, then retry."
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/BillingUnavailable"
          }
        },
        "summary": "Install a pack from the global registry",
        "tags": [
          "Domain Packs"
        ]
      }
    },
    "/v1/admin/packs/{packId}": {
      "delete": {
        "description": "Deprecates the pack's predicates; facts extracted with them survive.\n\nRequired scope: `brain:admin`.",
        "operationId": "uninstallDomainPack",
        "parameters": [
          {
            "description": "The installed pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UninstallPackResponse"
                }
              }
            },
            "description": "Uninstalled."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Uninstall a pack from this tenant",
        "tags": [
          "Domain Packs"
        ]
      }
    },
    "/v1/admin/packs/{packId}/eval": {
      "post": {
        "description": "Scores whether extraction (with the pack's predicates + extractionProfile active) still meets the pack’s own expectations. Runs up to MAX_EVAL_FIXTURES live LLM calls — throttled to 3 requests/min per credential.\n\nRequired scope: `brain:admin`.",
        "operationId": "evalDomainPack",
        "parameters": [
          {
            "description": "The installed pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "'union' (default) scores the shared union prompt; 'dedicated' runs the pack-scoped dedicated prompt instead.",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "enum": [
                "union",
                "dedicated"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackEvalReport"
                }
              }
            },
            "description": "The eval report."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "summary": "Run a pack's eval fixtures against the live extractor",
        "tags": [
          "Domain Packs"
        ]
      }
    },
...
    "/v1/admin/registry/packs/{packId}/checkout": {
      "post": {
        "description": "Creates a billing checkout session for the BUYING tenant (the caller's company is the billing userId). Open `checkoutUrl`, pay, then retry POST /v1/admin/packs/from-registry — the 402 hint on that route points here. Answers 400 for a free pack or while the billing integration is disabled. An `idempotency-key` header is forwarded to billing so client retries collapse into one order. Throttled to 10 requests/min per credential.\n\nRequired scope: `brain:admin`.",
        "operationId": "createRegistryPackCheckout",
        "parameters": [
          {
            "description": "The paid pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "The checkout session."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/BillingUnavailable"
          }
        },
        "summary": "Start a hosted checkout for a paid pack",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/v1/admin/registry/packs/{packId}/feature": {
      "post": {
        "description": "Surfaces the pack in the featured section on top of the catalogue listing and /registry/ui. Instance-local; never mirrored.\n\nRequired scope: `registry:curate`.",
        "operationId": "featureRegistryPack",
        "parameters": [
          {
            "description": "The pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureResponse"
                }
              }
            },
            "description": "The curation state."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Feature a pack (hosting-operator curation)",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/v1/admin/registry/packs/{packId}/pricing": {
      "delete": {
        "description": "Clears the paid flag — installs stop requiring an entitlement. Billing products/prices are immutable and stay.\n\nRequired scope: `registry:publish`.",
        "operationId": "clearRegistryPackPricing",
        "parameters": [
          {
            "description": "The pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackPricingResponse"
                }
              }
            },
            "description": "The pricing state."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Make a pack free again (publisher-owned)",
        "tags": [
          "Marketplace"
        ]
      },
      "put": {
        "description": "Marks the pack paid: ensures the billing product exists (entitlement key `domain_pack:<packId>`), mints a fresh immutable price, and stores the priceCode + display price in the instance-local registry meta. Only the company that published the pack may price it. Answers 400 while the billing integration is disabled. Source: src/registry/marketplace-admin.controller.ts.\n\nRequired scope: `registry:publish`.",
        "operationId": "setRegistryPackPricing",
        "parameters": [
          {
            "description": "The pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackPricingResponse"
                }
              }
            },
            "description": "The pricing state."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Price a pack (publisher-owned)",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/v1/admin/registry/packs/{packId}/unfeature": {
      "post": {
        "description": "Reverses a feature.\n\nRequired scope: `registry:curate`.",
        "operationId": "unfeatureRegistryPack",
        "parameters": [
          {
            "description": "The pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureResponse"
                }
              }
            },
            "description": "The curation state."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Remove a pack from the featured section",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/v1/admin/registry/packs/{packId}/{version}/unyank": {
      "post": {
        "description": "Reverses a yank — the version becomes installable again.\n\nRequired scope: `registry:publish`.",
        "operationId": "unyankRegistryPack",
        "parameters": [
          {
            "description": "The pack id.",
            "in": "path",
            "name": "packId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The yanked version to restore.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YankPackResponse"
                }
              }
            },
            "description": "Yank state after the call."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Restore a yanked version",
        "tags": [
          "Registry publishing"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Hosted instance (placeholder — self-hosted deployments serve the same paths on their own origin).",
      "url": "https://brain.inite.ai"
    }
  ],
...
}